Fully transitioned all* binaries to a proper arg parser! + Extras
Updated all binaries except sudo to use a custom version of python's argparser. Sudo is different because it has flags that doesn't specifically belong to the sudo command, but to the command sudo will execute. Also added list of packages that may or may not have a use in the future. May be useful for story mode.